chore: changed a look of aboutus component
