--- client/src/components/OnePage/AboutUs/AboutUsTexts/AboutUsTexts.tsx
@@ -1,25 +1,0 @@
-import React from 'react';
-
-const AboutUsTexts = () => {
-  return (
-    <div
-      className="container"
-      style={{
-        position: 'absolute',
-        justifyContent: 'center',
-        textAlign: 'center',
-        marginTop: '450px',
-        marginLeft: '80px',
-        color: 'white',
-      }}
-    >
-      <div className="text position-left">
-        <p>Surprising</p>
-        <p>and playful ways</p>
-        <p>to tell a story!</p>
-      </div>
-    </div>
-  );
-};
-
-export default AboutUsTexts;
