+++ client/src/components/OnePage/Banner.tsx
@@ -5,6 +5,7 @@ const BannerContainer = styled(Container)(({ theme }) => ({
   display: 'flex',
   flexDirection: 'column',
   position: 'relative',
+  justifyContent: 'center',
   width: '100%',
   height: '100vh',
   [theme.breakpoints.down('sm')]: {
